fixed the CreditAccount

+++ src/main/java/ru/netology/javaqadiplom/CreditAccount.java
@@ -80,7 +80,7 @@ public class CreditAccount extends Account {
      */
     @Override
     public int yearChange() {
-        return balance / 100 * rate;
+        return (int)(long)balance / 100 * rate;
     }
 
     public int getCreditLimit() {
